[M] Treat qbl_init as a sourceable init script
Installed into setup scripts dir by makefile
Only loaded once

+++ qbl_init.bash
@@ -3,11 +3,14 @@
 
 # Sets up coprocess for IPC, creates prompt maker, handles deletion of unix socket at process death, updates prompt
 # Trap is to avoid killing nc coproc when sending ctrl-C to the shell
-coproc CO_NC_GIT { trap -- "true" EXIT; nc -klU /tmp/.QBL_$$; } && disown $!
-if [[ -n "$CO_NC_GIT" ]]; then
-    git_lineutils $$ & disown $!
-    LINEUTILS_PID=$!
-    source bashline_colors
-    PROMPT_COMMAND="source q_bashline${PROMPT_COMMAND:+; $PROMPT_COMMAND}"
-    trap "rm -f /tmp/.QBL_$$ >/dev/null; kill $LINEUTILS_PID" EXIT
+if [[ "$QBL_LOADED" -ne '1' ]]; then
+    coproc CO_NC_GIT { trap -- "true" EXIT; nc -klU /tmp/.QBL_$$; } && disown $!
+    if [[ -n "$CO_NC_GIT" ]]; then
+        git_lineutils $$ & disown $!
+        LINEUTILS_PID=$!
+        source bashline_colors
+        PROMPT_COMMAND="source q_bashline${PROMPT_COMMAND:+; $PROMPT_COMMAND}"
+        trap "rm -f /tmp/.QBL_$$ >/dev/null; kill $LINEUTILS_PID" EXIT
+    fi
+    QBL_LOADED=1
 fi
